Delete the content of api/__init__.py and refactor the way of importing apis
